os detector updated
now changes to the dir before running iguana
checks if the directory already exists or not
if doesn’t creates the dir
changes the dir specific to os, and starts iguana from that dir.
found in previous updates iguana binaries was not included, or somehow
was missing. added to this update.